+++ src/classes/Application.php
@@ -73,6 +73,10 @@ class Application
 
     public const REQUEST_VAR_SIMULATION = 'simulate_only';
     public const REQUEST_VAR_QUERY_SUMMARY = 'query_summary';
+    public const STORAGE_FOLDER_NAME = 'storage';
+    public const TEMP_FOLDER_NAME = 'temp';
+    public const CACHE_FOLDER_NAME = 'cache';
+    public const DEFAULT_TEST_FILE_EXTENSION = 'tmp';
 
     private UI $ui;
     private ?Application_Driver $driver = null;
@@ -545,30 +549,48 @@ class Application
      */
     public static function getTempFolder() : string
     {
-        return self::getStorageSubfolderPath('temp');
+        return self::getStorageSubfolderPath(self::TEMP_FOLDER_NAME);
+    }
+
+    public static function getTempFolderURL() : string
+    {
+        return self::getStorageSubfolderURL(self::TEMP_FOLDER_NAME);
     }
 
     public static function getCacheFolder() : string
     {
-        return self::getStorageSubfolderPath('cache');
+        return self::getStorageSubfolderPath(self::CACHE_FOLDER_NAME);
     }
 
     /**
      * Generates a temporary file path.
      *
-     * Note: the file is not created. This just creates a file path.
+     * > Note: the file is not created. This just creates a file path.
      *
-     * @param string $name Specific name to use (without extension), or auto-generated if empty.
-     * @param string $extension The extension to use
+     * @param string|NULL $name Specific name to use (without extension), or auto-generated if empty.
+     * @param string|NULL $extension The extension to use, if empty {@see self::DEFAULT_TEST_FILE_EXTENSION} is used.
      */
-    public static function getTempFile(string $name = '', string $extension = 'tmp') : string
+    public static function getTempFile(?string $name = null, ?string $extension = self::DEFAULT_TEST_FILE_EXTENSION) : string
     {
-        if (empty($name))
-        {
+        return self::getTempFolder() . '/' . self::resolveTempFileName($name, $extension);
+    }
+
+    public static function getTempFileURL(?string $name = null, ?string $extension = self::DEFAULT_TEST_FILE_EXTENSION) : string
+    {
+        return self::getTempFolderURL() . '/' . self::resolveTempFileName($name, $extension);
+    }
+
+    private static function resolveTempFileName(?string $name = null, ?string $extension = self::DEFAULT_TEST_FILE_EXTENSION) : string
+    {
+        if (empty($name)) {
             $name = md5('tmp' . microtime(true));
         }
 
-        return self::getTempFolder() . '/' . $name . '.' . $extension;
+        if(empty($extension)) {
+            $extension = self::DEFAULT_TEST_FILE_EXTENSION;
+        }
+
+        return $name . '.' . $extension;
     }
 
     /**
@@ -585,7 +607,7 @@ class Application
             return self::$storageFolder;
         }
 
-        self::$storageFolder = APP_ROOT . '/storage';
+        self::$storageFolder = APP_ROOT . '/'. self::STORAGE_FOLDER_NAME;
 
         try
         {
@@ -660,6 +682,16 @@ class Application
         self::$knownStorageFolders[$subfolderName] = $folder;
 
         return $folder;
+    }
+
+    public static function getStorageSubfolderURL(string $subfolderName) : string
+    {
+        return sprintf(
+            '%s/%s/%s',
+            APP_URL,
+            self::STORAGE_FOLDER_NAME,
+            $subfolderName
+        );
     }
 
     /**